Move Flask from /app to /api
