+++ api/gumbaroo.py
@@ -1,4 +1,4 @@
-from app import app
+from api import app
 from flask import jsonify
